backend/agent/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.2
langchain>=0.3.6
langgraph>=0.2.34
langchain-openai>=0.2.5
langsmith>=0.1.79
pymongo>=4.8.0
tiktoken>=0.7.0
rich>=13.8.1
numpy>=1.26.4
sentence-transformers>=3.0.1
scikit-learn>=1.5.2
pymongo[aws]
motor>=3.5.1
fastapi>=0.115.2
uvicorn>=0.31.1
langchain-community>=0.3.0
python-json-logger>=2.0.7
boto3>=1.35.0
opensearch-py>=2.3.0
requests-aws4auth>=1.2.3
spacy>=3.8.0
presidio-analyzer>=2.2.355
presidio-anonymizer>=2.2.355
sentencepiece>=0.1.99
accelerate>=0.27.0